split rest-api part 3
