+++ src/components/diagrams/singleLineDiagram/single-line-diagram-content.js
@@ -29,10 +29,7 @@ import withBranchMenu from '../../menus/branch-menu';
 import { SingleLineDiagramViewer } from '@powsybl/diagram-viewer';
 import { isNodeReadOnly } from '../../graph/util/model-functions';
 import { useIsAnyNodeBuilding } from '../../utils/is-any-node-building-hook';
-import {
-    startShortCircuitAnalysis,
-    fetchNetworkElementInfos,
-} from '../../../utils/rest-api';
+import { fetchNetworkElementInfos } from '../../../utils/rest-api';
 import Alert from '@mui/material/Alert';
 import { useTheme } from '@mui/material/styles';
 import { useSnackMessage } from '@gridsuite/commons-ui';
@@ -58,6 +55,7 @@ import {
     EQUIPMENT_TYPES,
 } from '../../utils/equipment-types';
 import EquipmentDeletionDialog from '../../dialogs/network-modifications/equipment-deletion/equipment-deletion-dialog';
+import { startShortCircuitAnalysis } from '../../../services/study/short-circuit-analysis';
 
 function SingleLineDiagramContent(props) {
     const { studyUuid } = props;
