+++ src/components/dialogs/network-modifications/battery/creation/battery-creation-dialog.js
@@ -40,7 +40,7 @@ import {
     getConnectivityWithPositionValidationSchema,
 } from '../../../connectivity/connectivity-form-utils';
 import BatteryCreationForm from './battery-creation-form';
-import { createBattery, FetchStatus } from 'utils/rest-api';
+import { FetchStatus } from 'utils/rest-api';
 import { sanitizeString } from '../../../dialogUtils';
 import {
     FORM_LOADING_DELAY,
@@ -54,6 +54,7 @@ import { useOpenShortWaitFetching } from 'components/dialogs/commons/handle-modi
 import { getFrequencyRegulationSchema } from '../../generator/set-points/set-points-utils';
 import { EQUIPMENT_TYPES } from '../../../../utils/equipment-types';
 import PropTypes from 'prop-types';
+import { createBattery } from '../../../../../services/study/network-modifications';
 
 const emptyFormData = {
     [EQUIPMENT_ID]: '',
